[REFACTOR] 물품 상세정보 훅 분리
